getting owner from githubClient, enforcing required env vars, and added retries to repository deletion attempts

+++ src/main/groovy/com/wooga/spock/extensions/github/Repository.groovy
@@ -32,6 +32,7 @@ class Repository {
 
     private interface GHRepositoryDelegateExcludes {
         String getDefaultBranch()
+        void delete()
     }
 
     @Delegate(excludeTypes = [GHRepositoryDelegateExcludes.class])
@@ -61,9 +62,9 @@ class Repository {
 
     boolean exists() {
         try {
-            client.getRepository(fullName)
+            client.getRepositoryById(repository.id)
             return true
-        } catch (ignored) {
+        } catch (FileNotFoundException) {
             return false
         }
     }
@@ -129,8 +130,21 @@ class Repository {
         findTag(tagName).orElse(null)
     }
 
-    void delete() {
-        repository.delete()
+    void delete(int retries=3) {
+        try {
+            if(this.exists()) {
+                repository.delete()
+            }
+        } catch (FileNotFoundException _) {
+        } catch (Exception e) {
+            if(retries > 0) {
+                sleep(500)
+                delete(retries-1)
+            } else {
+                throw e
+            }
+        }
+
     }
 
     Optional<GHTag> findTag(String tagName) {
